# Improve coalescing performance by using a FxHashMap