This improves the performance of `VirtualCopies::filter` by using a `FxHashMap` instead of a binary search.

I also changed all the `HashMap` in the `codegen` crate to `FxHashMap` for consistency, but there is no noticeable performance improvement since 80% of the time is spent in the register allocator.

## Performance results

Result from compiling `sqlite.wasm` show a 16.7% reduction in coalescing time, which translates to a 9.6% reduction in total compilation time.

Before:
```
$ perf stat target/release/cton-util wasm -T --set enable_verifier=0 --set is_64bit=1 --isa x86 ../wasm-collection/misc-valid/sqlite.wasm
======== ========  ==================================
   Total     Self  Pass
-------- --------  ----------------------------------
   0.319    0.001  Translate WASM module
   0.318    0.318  Translate WASM function
   2.961    0.037  Compilation passes
   0.009    0.009  Control flow graph
   0.014    0.014  Dominator tree
   0.034    0.032  Post-legalization rewriting
   0.017    0.017  Pre-legalization rewriting
   0.018    0.018  Dead code elimination
   0.108    0.104  Legalization
   0.000    0.000  Remove unreachable blocks
   2.710    0.001  Register allocation
   0.276    0.276  RA liveness analysis
   1.722    1.708  RA coalescing CSSA
   0.241    0.241  RA spilling
   0.207    0.207  RA reloading
   0.264    0.264  RA coloring
   0.013    0.013  Prologue/epilogue insertion
   0.020    0.020  Layout full renumbering
======== ========  ==================================

 Performance counter stats for 'target/release/cton-util wasm -T --set enable_verifier=0 --set is_64bit=1 --isa x86 ../wasm-collection/misc-valid/sqlite.wasm':

       3331.796606      task-clock (msec)         #    1.000 CPUs utilized          
                22      context-switches          #    0.007 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            60,273      page-faults               #    0.018 M/sec                  
    10,049,185,059      cycles                    #    3.016 GHz                    
    22,985,804,643      instructions              #    2.29  insn per cycle         
     3,812,377,156      branches                  # 1144.241 M/sec                  
        54,570,693      branch-misses             #    1.43% of all branches        

       3.332527015 seconds time elapsed
```

After:
```
$ perf stat target/release/cton-util wasm -T --set enable_verifier=0 --set is_64bit=1 --isa x86 ../wasm-collection/misc-valid/sqlite.wasm
======== ========  ==================================
   Total     Self  Pass
-------- --------  ----------------------------------
   0.312    0.001  Translate WASM module
   0.311    0.311  Translate WASM function
   2.699    0.037  Compilation passes
   0.009    0.009  Control flow graph
   0.015    0.015  Dominator tree
   0.034    0.032  Post-legalization rewriting
   0.017    0.017  Pre-legalization rewriting
   0.017    0.017  Dead code elimination
   0.102    0.098  Legalization
   0.000    0.000  Remove unreachable blocks
   2.456    0.001  Register allocation
   0.292    0.292  RA liveness analysis
   1.475    1.459  RA coalescing CSSA
   0.233    0.233  RA spilling
   0.196    0.196  RA reloading
   0.259    0.259  RA coloring
   0.012    0.012  Prologue/epilogue insertion
   0.022    0.022  Layout full renumbering
======== ========  ==================================

 Performance counter stats for 'target/release/cton-util wasm -T --set enable_verifier=0 --set is_64bit=1 --isa x86 ../wasm-collection/misc-valid/sqlite.wasm':

       3062.545874      task-clock (msec)         #    1.000 CPUs utilized          
                 7      context-switches          #    0.002 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            58,830      page-faults               #    0.019 M/sec                  
     9,542,870,453      cycles                    #    3.116 GHz                    
    21,550,752,471      instructions              #    2.26  insn per cycle         
     3,458,311,590      branches                  # 1129.228 M/sec                  
        48,707,280      branch-misses             #    1.41% of all branches        

       3.063093766 seconds time elapsed
```