@Amanieu 2018-05-17
Open question: should we also use `FxHashMap` for the other cretonne crates?