@sunfishcode 2018-05-17
> Open question: should we also use FxHashMap for the other cretonne crates?

If we arrange for the name in the code to be `HashMap`, then yes, I'd be inclined to use it everywhere. Most of the HashMaps in the other crates are probably not performance-sensitive, but some of the ones in cretonne-wasm may be in some situations.