@sunfishcode 2018-05-17
> I copied the implementation from rustc instead of from the fxhash crate. I haven't made any changes to the code myself.

Ah, cool. In that case, please add a comment somewhere mentioning the upstream source for this file.